# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ticket Analyzer AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets/uploads/{id}:
    delete:
      operationId: deleteUpload
      tags: [tickets]
      summary: Delete an upload and its tickets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /tickets/dashboard:
    get:
      operationId: getDashboard
      tags: [tickets]
      summary: Get dashboard statistics
      parameters:
        - name: upload_id
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /tickets/list:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets with optional filters
      parameters:
        - name: upload_id
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: assignee
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    Ticket:
      type: object
      required: [id, upload_id]
      properties:
        id:
          type: integer
        upload_id:
          type: integer
        ticket_id:
          type: ["string", "null"]
        ticket_date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        assignee:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]
        ai_category:
          type: ["string", "null"]
        is_anomaly:
          type: ["boolean", "null"]
    ChartDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: integer
    DashboardData:
      type: object
      required: [total, byCategory, byPriority, byStatus, byAssignee, byWeek]
      properties:
        total:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        byPriority:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        byAssignee:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        byWeek:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        openP1Count:
          type: integer
        avgResolutionDays:
          type: ["number", "null"]
...